Runtime: Fix unsigned/signed warning

+++ src/libb0rk/runtime.cpp
@@ -226,7 +226,7 @@ Class* Runtime::loadClass(Context* context, wstring name, bool addToExisting)
 
     char* buffer = new char[length + 128];
     memset(buffer + length, 0, 128);
-    int read = fread(buffer, 1, length, fp);
+    size_t read = fread(buffer, 1, length, fp);
     fclose(fp);
 
     if (read != length)
